actualización código para que pueda enviar correctamente en formato json

+++ bd.sql
@@ -3,31 +3,31 @@ create database tareas;
 use tareas;
 
 create table proyecto(
-idproyecto int auto_increment primary key,
+idproyecto bigint auto_increment primary key,
 proyecto varchar(200)
 );
 
 create table tipo_doc(
-idtipo_doc int auto_increment primary key,
+idtipo_doc bigint auto_increment primary key,
 tipo_doc varchar(12)
 );
 
 create table persona(
-idpersona int auto_increment primary key,
+idpersona bigint auto_increment primary key,
 documento varchar(12),
-tipo_doc int,
+tipo_doc bigint,
 foreign key (tipo_doc) references tipo_doc(idtipo_doc),
 nombre varchar(50),
 apellido varchar(50)
 );
 
 create table tareas(
-idtareas int auto_increment primary key,
+idtareas bigint auto_increment primary key,
 tarea varchar(100),
 descripción text,
 fecha_inicio date,
-idproyecto int,
+idproyecto bigint,
 foreign key (idproyecto) references proyecto(idproyecto),
-idpersona int,
+idpersona bigint,
 foreign key (idpersona) references persona (idpersona) 
 );
